deepagents/projects/project-3.12-workflow-recovery-agent/requirements.txt
langchain-core>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.1
